daemon/requirements.txt
adafruit-circuitpython-dht
fastapi-restful
uvicorn[standard]
python-dotenv
asyncio
asyncpg
